System.Statgrab

Description

Monadic context and data types for managing the underlying libstatgrab FFI calls with transparent resource allocation and deallocation.

Running the Stats Monad

data Stats a

runStats :: MonadCatchIO m => Stats a -> m a

Run the Stats Monad, bracketing libstatgrab's sg_init and sg_shutdown calls via reference counting to ensure reentrancy.

async :: Stats a -> Stats (Async a)

Run the Stats Monad asynchronously. wait from the async package can be used to block and retrieve the result of the asynchronous computation.

Retrieving Statistics

snapshot :: (Stat (Struct a), Copy a) => Stats a

Retrieve statistics from the underlying operating system, copying them to the Haskell heap and freeing the related Ptr a.

The *_r variants of the libstatgrab functions are used and the deallocation strategy is bracketed.

class Stat a

data family Struct a

Statistic Types

data Host

Constructors

Host

Fields

hostOsName :: !ByteString

hostOsRelease :: !ByteString

hostOsVersion :: !ByteString

hostPlatform :: !ByteString

hostName :: !ByteString

hostBitWidth :: !Integer

hostState :: !HostState

hostNCPU :: !Integer

hostMaxCPU :: !Integer

hostUptime :: !POSIXTime

hostSystime :: !POSIXTime

data CPU

Constructors

CPU

Fields

cpuUser :: !Integer

cpuKernel :: !Integer

cpuIdle :: !Integer

cpuIOWait :: !Integer

cpuSwap :: !Integer

cpuNice :: !Integer

cpuTotal :: !Integer

cpuCtxSwitches :: !Integer

cpuVoluntaryCtxSwitches :: !Integer

cpuInvoluntaryCtxSwitches :: !Integer

cpuSyscalls :: !Integer

cpuInterrupts :: !Integer

cpuSoftInterrupts :: !Integer

cpuSystime :: !POSIXTime

data CPUPercent

Constructors

CPUPercent

Fields

cpuPctUser :: !Double

cpuPctKernel :: !Double

cpuPctIdle :: !Double

cpuPctIOWait :: !Double

cpuPctSwap :: !Double

cpuPctNice :: !Double

cpuPctTimeTaken :: !POSIXTime

data Memory

Constructors

Memory

Fields

memTotal :: !Integer

memFree :: !Integer

memUsed :: !Integer

memCache :: !Integer

memSystime :: !POSIXTime

data Load

Constructors

Load

Fields

load1 :: !Double

load5 :: !Double

load15 :: !Double

loadSystime :: !POSIXTime

data User

Constructors

User

Fields

userLoginName :: !ByteString

userRecordId :: !ByteString

userRecordIdSize :: !Integer

userDevice :: !ByteString

userHostName :: !ByteString

userPid :: !Integer

userLoginTime :: !POSIXTime

userSystime :: !POSIXTime

data Swap

Constructors

Swap

Fields

swapTotal :: !Integer

swapUsed :: !Integer

swapFree :: !Integer

swapSystime :: !POSIXTime

data FileSystem

Constructors

FileSystem

Fields

fsDeviceName :: !ByteString

fsType :: !ByteString

fsMountPoint :: !ByteString

fsDeviceType :: !DeviceType

fsSize :: !Integer

fsUsed :: !Integer

fsFree :: !Integer

fsAvail :: !Integer

fsTotalInodes :: !Integer

fsUsedInodes :: !Integer

fsFreeInodes :: !Integer

fsAvailInodes :: !Integer

fsIOSize :: !Integer

fsBlockSize :: !Integer

fsTotalBlocks :: !Integer

fsFreeBlocks :: !Integer

fsUsedBlocks :: !Integer

fsAvailBlocks :: !Integer

fsSystime :: !POSIXTime

data DiskIO

Constructors

DiskIO

Fields

diskName :: !ByteString

diskRead :: !Integer

diskWrite :: !Integer

diskSystime :: !POSIXTime

data NetworkIO

Constructors

NetworkIO

Fields

ifaceIOName :: !ByteString

ifaceTX :: !Integer

ifaceRX :: !Integer

ifaceIPackets :: !Integer

ifaceOPackets :: !Integer

ifaceIErrors :: !Integer

ifaceOErrors :: !Integer

ifaceCollisions :: !Integer

ifaceSystem :: !POSIXTime

data NetworkInterface

Constructors

NetworkInterface

Fields

ifaceName :: !ByteString

ifaceSpeed :: !Integer

ifaceFactor :: !Integer

ifaceDuplex :: !InterfaceMode

ifaceUp :: !InterfaceStatus

ifaceSystime :: !POSIXTime

data Page

Constructors

Page

Fields

pagesIn :: !Integer

pagesOut :: !Integer

pagesSysTime :: !POSIXTime

data Process

Constructors

Process

Fields

procName :: !ByteString

procTitle :: !ByteString

procPid :: !Integer

procParent :: !Integer

procPGid :: !Integer

procSessId :: !Integer

procUid :: !Integer

procEUid :: !Integer

procGid :: !Integer

procEGid :: !Integer

procSwitches :: !Integer

procVoluntary :: !Integer

procInvoluntary :: !Integer

procSize :: !Integer

procResident :: !Integer

procStart :: !POSIXTime

procSpent :: !POSIXTime

procCPUPercent :: !Double

procNice :: !Integer

procState :: !ProcessState

procSystime :: !POSIXTime

data ProcessCount

Constructors

ProcessCount

Fields

countTotal :: !Integer

countRunning :: !Integer

countSleeping :: !Integer

countStopped :: !Integer

countZombie :: !Integer

countUnknown :: !Integer

countSystime :: !POSIXTime

Enums

newtype HostState

Constructors

HostState CInt

newtype CPUPercentSource

Constructors

CPUPercentSource CInt

newtype DeviceType

Constructors

DeviceType CInt

newtype InterfaceMode

Constructors

InterfaceMode CInt

newtype InterfaceStatus

Constructors

InterfaceStatus CInt

newtype ProcessState

Constructors

ProcessState CInt

newtype ProcessSource

Constructors

ProcessSource CInt

Re-exported

data Async a

wait :: Async a -> IO a